Move applySecuritySettings() back into CmdLineTool
This allows us to share the code with the forwarder
