+++ graylog2-server/src/main/java/org/graylog2/bootstrap/ServerBootstrap.java
@@ -22,7 +22,6 @@ import com.google.common.util.concurrent.ServiceManager;
 import com.google.inject.Injector;
 import com.google.inject.Module;
 import com.google.inject.ProvisionException;
-import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.graylog2.audit.AuditActor;
 import org.graylog2.audit.AuditEventSender;
 import org.graylog2.configuration.TLSProtocolsConfiguration;
@@ -52,10 +51,8 @@ import java.nio.file.LinkOption;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
-import java.security.Security;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
@@ -101,31 +98,6 @@ public abstract class ServerBootstrap extends CmdLineTool {
 
         // Set these early in the startup because netty's NativeLibraryUtil uses a static initializer
         setNettyNativeDefaults();
-    }
-
-    private static void applySecuritySettings(TLSProtocolsConfiguration configuration) {
-        // Disable insecure TLS parameters and algorithms by default.
-        // Prevent attacks like LOGJAM, LUCKY13, et al.
-        setSystemPropertyIfEmpty("jdk.tls.ephemeralDHKeySize", "2048");
-        setSystemPropertyIfEmpty("jdk.tls.rejectClientInitiatedRenegotiation", "true");
-
-        // Only restrict ciphers if insecure TLS protocols are explicitly enabled.
-        // c.f. https://github.com/Graylog2/graylog2-server/issues/10944
-        final Set<String> tlsProtocols = configuration.getConfiguredTlsProtocols();
-        if (tlsProtocols == null || !(tlsProtocols.isEmpty() || tlsProtocols.contains("TLSv1") || tlsProtocols.contains("TLSv1.1") || tlsProtocols.contains("TLSv1.0"))) {
-            // Weirdly this is not a System property
-            Security.setProperty("jdk.tls.disabledAlgorithms", "CBC,3DES");
-        }
-
-        // Explicitly register Bouncy Castle as security provider.
-        // This allows us to use more key formats than with JCE
-        Security.addProvider(new BouncyCastleProvider());
-    }
-
-    private static void setSystemPropertyIfEmpty(String key, String value) {
-        if (System.getProperty(key) == null) {
-            System.setProperty(key, value);
-        }
     }
 
     private void setNettyNativeDefaults() {
